Commonize and improve clock frequency conversion.
- Moved convert_frequency() to pyocd.utility.cmdline.
- Added support for float values and 'Hz' suffix.
- Using convert_frequency() for 'set clock' in commander.
