+++ pyocd/__main__.py
@@ -41,6 +41,7 @@ from .utility.cmdline import (
     convert_vector_catch,
     convert_session_options,
     convert_reset_type,
+    convert_frequency,
     )
 from .probe.pydapaccess import DAPAccess
 from .tools.lists import ListGenerator
@@ -81,20 +82,6 @@ ERASE_OPTIONS = [
     'chip',
     'sector',
     ]
-
-def convert_frequency(value):
-    """! @brief Applies scale suffix to frequency value string."""
-    value = value.strip()
-    suffix = value[-1].lower()
-    if suffix in ('k', 'm'):
-        value = int(value[:-1])
-        if suffix == 'k':
-            value *= 1000
-        elif suffix == 'm':
-            value *= 1000000
-        return value
-    else:
-        return int(value)
 
 def flatten_args(args):
     """! @brief Converts a list of lists to a single list."""
@@ -164,7 +151,8 @@ class PyOCDTool(object):
         connectOptions.add_argument("-t", "--target", dest="target_override", metavar="TARGET",
             help="Set the target type.")
         connectOptions.add_argument("-f", "--frequency", dest="frequency", default=None, type=convert_frequency,
-            help="SWD/JTAG clock frequency in Hz, with optional k/K or m/M suffix for kHz or MHz.")
+            help="SWD/JTAG clock frequency in Hz. Accepts a float or int with optional case-"
+                "insensitive K/M suffix and optional Hz. Examples: \"1000\", \"2.5khz\", \"10m\".")
         connectOptions.add_argument("-W", "--no-wait", action="store_true",
             help="Do not wait for a probe to be connected if none are available.")
         connectOptions.add_argument("-M", "--connect", dest="connect_mode", metavar="MODE",
